# [GEOS-11284] Promote community module "datadir catalog loader" to core