--- a/doc/en/user/source/community/index.rst
+++ b/doc/en/user/source/community/index.rst
@@ -77,4 +77,3 @@ officially part of the GeoServer releases. They are however built along with the
    webp/index
    wfs-freemarker/index
    wps-longitudinal-profile/index
-   datadir-catalog-loader/index