--- a/src/main/src/main/java/org/geoserver/catalog/impl/AbstractCatalogFacade.java
+++ b/src/main/src/main/java/org/geoserver/catalog/impl/AbstractCatalogFacade.java
@@ -7,27 +7,20 @@ package org.geoserver.catalog.impl;
 
 import java.lang.reflect.Proxy;
 import java.rmi.server.UID;
-import java.util.LinkedHashSet;
 import java.util.List;
-import java.util.logging.Level;
-import java.util.logging.Logger;
 import org.geoserver.catalog.CatalogFacade;
 import org.geoserver.catalog.CatalogInfo;
 import org.geoserver.catalog.LayerGroupInfo;
 import org.geoserver.catalog.LayerInfo;
 import org.geoserver.catalog.MapInfo;
 import org.geoserver.catalog.NamespaceInfo;
-import org.geoserver.catalog.PublishedInfo;
 import org.geoserver.catalog.ResourceInfo;
 import org.geoserver.catalog.StoreInfo;
 import org.geoserver.catalog.StyleInfo;
 import org.geoserver.catalog.WorkspaceInfo;
 import org.geoserver.ows.util.OwsUtils;
-import org.geotools.util.logging.Logging;
 
 public abstract class AbstractCatalogFacade implements CatalogFacade {
-
-    private static final Logger LOGGER = Logging.getLogger(AbstractCatalogFacade.class);
 
     //
     // Utilities
@@ -68,167 +61,42 @@ public abstract class AbstractCatalogFacade implements CatalogFacade {
 
     protected void resolve(LayerInfo layer) {
         setId(layer);
-
-        ResourceInfo resource = ResolvingProxy.resolve(getCatalog(), layer.getResource());
-        if (resource != null) {
-            resource = unwrap(resource);
-            layer.setResource(resource);
-        }
-
-        StyleInfo style = ResolvingProxy.resolve(getCatalog(), layer.getDefaultStyle());
-        if (style != null) {
-            style = unwrap(style);
-            layer.setDefaultStyle(style);
-        }
-
-        LinkedHashSet<StyleInfo> styles = new LinkedHashSet<>();
-        for (StyleInfo s : layer.getStyles()) {
-            s = ResolvingProxy.resolve(getCatalog(), s);
-            s = unwrap(s);
-            styles.add(s);
-        }
-        ((LayerInfoImpl) layer).setStyles(styles);
+        ResolvingProxyResolver.resolve(layer, getCatalog());
     }
 
     protected void resolve(LayerGroupInfo layerGroup) {
         setId(layerGroup);
-
-        LayerGroupInfoImpl lg = (LayerGroupInfoImpl) layerGroup;
-
-        resolveLayerGroupLayers(lg.getLayers());
-        resolveLayerGroupStyles(lg.getLayers(), lg.getStyles());
-        // now resolves layers and styles defined in layer group styles
-        for (LayerGroupStyle groupStyle : lg.getLayerGroupStyles()) {
-            resolveLayerGroupLayers(groupStyle.getLayers());
-            resolveLayerGroupStyles(groupStyle.getLayers(), groupStyle.getStyles());
-        }
-    }
-
-    private void resolveLayerGroupStyles(List<PublishedInfo> assignedLayers, List<StyleInfo> styles) {
-        for (int i = 0; i < styles.size(); i++) {
-            StyleInfo s = styles.get(i);
-            if (s != null) {
-                PublishedInfo assignedLayer = assignedLayers.get(i);
-                StyleInfo resolved = null;
-                if (assignedLayer instanceof LayerGroupInfo) {
-                    // special case we might have a StyleInfo representing
-                    // only the name of a LayerGroupStyle thus not present in Catalog.
-                    // We take the ref and create a new object
-                    // without searching in catalog.
-                    String ref = ResolvingProxy.getRef(s);
-                    if (ref != null) {
-                        StyleInfo styleInfo = new StyleInfoImpl(getCatalog());
-                        styleInfo.setName(ref);
-                        resolved = styleInfo;
-                    }
-                }
-                if (resolved == null) resolved = unwrap(ResolvingProxy.resolve(getCatalog(), s));
-
-                styles.set(i, resolved);
-            }
-        }
-    }
-
-    private void resolveLayerGroupLayers(List<PublishedInfo> layers) {
-        for (int i = 0; i < layers.size(); i++) {
-            PublishedInfo l = layers.get(i);
-
-            if (l != null) {
-                PublishedInfo resolved;
-                if (l instanceof LayerGroupInfo) {
-                    resolved = unwrap(ResolvingProxy.resolve(getCatalog(), (LayerGroupInfo) l));
-                    // special case to handle catalog loading, when nested publishibles might not be
-                    // loaded.
-                    if (resolved == null) {
-                        resolved = l;
-                    }
-                } else if (l instanceof LayerInfo) {
-                    resolved = unwrap(ResolvingProxy.resolve(getCatalog(), (LayerInfo) l));
-                    // special case to handle catalog loading, when nested publishibles might not be
-                    // loaded.
-                    if (resolved == null) {
-                        resolved = l;
-                    }
-                } else {
-                    // Special case for null layer (style group)
-                    resolved = unwrap(ResolvingProxy.resolve(getCatalog(), l));
-                }
-                layers.set(i, resolved);
-            }
-        }
+        ResolvingProxyResolver.resolve(layerGroup, getCatalog());
     }
 
     protected void resolve(StyleInfo style) {
         setId(style);
-        StyleInfoImpl s = (StyleInfoImpl) style;
-        s.setCatalog(getCatalog());
-
-        // resolve the workspace
-        WorkspaceInfo ws = s.getWorkspace();
-        if (ws != null) {
-            WorkspaceInfo resolved = ResolvingProxy.resolve(getCatalog(), ws);
-            if (resolved != null) {
-                resolved = unwrap(resolved);
-                s.setWorkspace(resolved);
-            } else {
-                LOGGER.log(
-                        Level.INFO,
-                        "Failed to resolve workspace for style \""
-                                + style.getName()
-                                + "\". This means the workspace has not yet been added to the catalog, keep the proxy around");
-            }
-        }
+        ResolvingProxyResolver.resolve(style, getCatalog());
     }
 
     protected void resolve(MapInfo map) {
         setId(map);
+        ResolvingProxyResolver.resolve(map, getCatalog());
     }
 
     protected void resolve(WorkspaceInfo workspace) {
         setId(workspace);
+        ResolvingProxyResolver.resolve(workspace, getCatalog());
     }
 
     protected void resolve(NamespaceInfo namespace) {
         setId(namespace);
+        ResolvingProxyResolver.resolve(namespace, getCatalog());
     }
 
     protected void resolve(StoreInfo store) {
         setId(store);
-        StoreInfoImpl s = (StoreInfoImpl) store;
-        s.setCatalog(getCatalog());
-
-        // resolve the workspace
-        WorkspaceInfo resolved = ResolvingProxy.resolve(getCatalog(), s.getWorkspace());
-        if (resolved != null) {
-            resolved = unwrap(resolved);
-            s.setWorkspace(resolved);
-        } else {
-            LOGGER.log(
-                    Level.INFO,
-                    "Failed to resolve workspace for store \""
-                            + store.getName()
-                            + "\". This means the workspace has not yet been added to the catalog, keep the proxy around");
-        }
+        ResolvingProxyResolver.resolve(store, getCatalog());
     }
 
     protected void resolve(ResourceInfo resource) {
         setId(resource);
-        ResourceInfoImpl r = (ResourceInfoImpl) resource;
-        r.setCatalog(getCatalog());
-
-        // resolve the store
-        StoreInfo store = ResolvingProxy.resolve(getCatalog(), r.getStore());
-        if (store != null) {
-            store = unwrap(store);
-            r.setStore(store);
-        }
-
-        // resolve the namespace
-        NamespaceInfo namespace = ResolvingProxy.resolve(getCatalog(), r.getNamespace());
-        if (namespace != null) {
-            namespace = unwrap(namespace);
-            r.setNamespace(namespace);
-        }
+        ResolvingProxyResolver.resolve(resource, getCatalog());
     }
 
     protected void setId(Object o) {