--- a/src/main/src/main/java/org/geoserver/catalog/impl/NamespaceInfoLookup.java
+++ b/src/main/src/main/java/org/geoserver/catalog/impl/NamespaceInfoLookup.java
@@ -11,8 +11,6 @@ import java.util.Comparator;
 import java.util.List;
 import java.util.Objects;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.locks.ReadWriteLock;
-import java.util.concurrent.locks.ReentrantReadWriteLock;
 import org.geoserver.catalog.Catalog;
 import org.geoserver.catalog.NamespaceInfo;
 
@@ -21,13 +19,12 @@ import org.geoserver.catalog.NamespaceInfo;
  * {@link #findAllByUri(String) fast lookup} of namespaces by uri.
  *
  * <p>All {@link CatalogInfoLookup} mutating methods are overridden to maintain the index consistency
+ *
+ * @implNote this class uses {@link CatalogInfoLookup#lock} to guard modifications to the index and its ArrayList values
  */
 class NamespaceInfoLookup extends CatalogInfoLookup<NamespaceInfo> {
 
     private ConcurrentHashMap<String, List<NamespaceInfo>> index = new ConcurrentHashMap<>();
-
-    /** guards modifications to the index and its ArrayList values */
-    private ReadWriteLock lock = new ReentrantReadWriteLock();
 
     private static Comparator<NamespaceInfo> VALUE_ORDER =
             (n1, n2) -> n1.getId().compareTo(n2.getId());
@@ -55,13 +52,13 @@ class NamespaceInfoLookup extends CatalogInfoLookup<NamespaceInfo> {
 
     @Override
     public NamespaceInfo add(NamespaceInfo value) {
-        lock.writeLock().lock();
+        writeLock.lock();
         try {
             NamespaceInfo ns = super.add(value);
             addInternal(value);
             return ns;
         } finally {
-            lock.writeLock().unlock();
+            writeLock.unlock();
         }
     }
 
@@ -73,25 +70,25 @@ class NamespaceInfoLookup extends CatalogInfoLookup<NamespaceInfo> {
 
     @Override
     public void clear() {
-        lock.writeLock().lock();
+        writeLock.lock();
         try {
             super.clear();
             index.clear();
         } finally {
-            lock.writeLock().unlock();
+            writeLock.unlock();
         }
     }
 
     @Override
     public NamespaceInfo remove(NamespaceInfo value) {
-        lock.writeLock().lock();
+        writeLock.lock();
         try {
             String uri = value.getURI();
             NamespaceInfo ns = super.remove(value);
             removeInternal(value, uri);
             return ns;
         } finally {
-            lock.writeLock().unlock();
+            writeLock.unlock();
         }
     }
 
@@ -105,7 +102,7 @@ class NamespaceInfoLookup extends CatalogInfoLookup<NamespaceInfo> {
 
     @Override
     public void update(NamespaceInfo value) {
-        lock.writeLock().lock();
+        writeLock.lock();
         try {
             ModificationProxy h = (ModificationProxy) Proxy.getInvocationHandler(value);
             NamespaceInfo actualValue = (NamespaceInfo) h.getProxyObject();
@@ -121,7 +118,7 @@ class NamespaceInfoLookup extends CatalogInfoLookup<NamespaceInfo> {
                 addInternal(value);
             }
         } finally {
-            lock.writeLock().unlock();
+            writeLock.unlock();
         }
     }