… doc/en/user/source/community/datadir-catalog-loader/index.rst
… doc/en/user/source/datadirectory/setting.rst
… doc/en/user/source/installation/upgrade.rst
… src/community/datadir-catalog-loader/README.md
… src/community/datadir-catalog-loader/pom.xml
… src/community/datadir-catalog-loader/src/assembly/assembly.xml
… src/community/datadir-catalog-loader/src/main/java/org/geoserver/config/datadir/CatalogLoader.java
… src/community/datadir-catalog-loader/src/main/java/org/geoserver/config/datadir/config/DataDirectoryGeoServerLoaderConfiguration.java
… src/community/datadir-catalog-loader/src/main/resources/applicationContext.xml
… src/community/pom.xml
… src/main/pom.xml
… src/main/src/main/java/applicationContext.xml
… src/main/src/main/java/org/geoserver/catalog/impl/ResolvingProxyResolver.java
… src/main/src/main/java/org/geoserver/config/GeoServerLoader.java
… src/main/src/main/java/org/geoserver/config/GeoServerLoaderProxy.java
… src/main/src/main/java/org/geoserver/config/datadir/CatalogLoader.java
… src/main/src/main/java/org/geoserver/config/datadir/CatalogLoaderSanitizer.java
… src/main/src/main/java/org/geoserver/config/datadir/ConfigLoader.java
… src/main/src/main/java/org/geoserver/config/datadir/DataDirectoryGeoServerLoader.java
… src/main/src/main/java/org/geoserver/config/datadir/DataDirectoryWalker.java
… src/main/src/main/java/org/geoserver/config/datadir/ExecutorFactory.java
… src/main/src/main/java/org/geoserver/config/datadir/XStreamLoader.java
… src/main/src/main/resources/DEFAULT_LOGGING.xml
… src/main/src/main/resources/PRODUCTION_LOGGING.xml
… src/main/src/test/java/org/geoserver/catalog/impl/ResolvingProxyResolverTest.java
… src/main/src/test/java/org/geoserver/config/datadir/DataDirectoryGeoServerLoaderConfigurationTest.java
… src/main/src/test/java/org/geoserver/config/datadir/DataDirectoryGeoServerLoaderTest.java
… src/main/src/test/java/org/geoserver/config/datadir/DataDirectoryLoaderTestSupport.java
… src/main/src/test/java/org/geoserver/config/datadir/DefaultGeoServerLoaderCompatibilityTest.java
… src/pom.xml