@groldan 2025-03-06
@aaime and others, the old PR was a mess to rebase after the Palantir formatting.
Also, curiously enough, your [sample datadir](https://www.dropbox.com/scl/fi/ip3x82pdzui1rxnvwsqmb/large_shapefiles.tar.bz2?rlkey=7uk600hlcp9z6tdbsdka86dpl&st=bkz2knvy&dl=1), worked with the community module as it was, but not with the refactored code from the old pr.

In any case, I've cleaned things up a lot in this one, making the new loader play nicely with GeoServerLoader and DefaultGeoServerLoader, instead of re-doing a lot of their workflow.

Added a bunch of test cases for invalid datadir configs.

The next commit will be about removing the community module and moving these classes to core.

Meanwhile, any testing is very much welcomed.

You can download these two files to run GeoServer against the sample data directory from docker compose:
* [initialize_volume.sh](https://gist.github.com/groldan/b09df3377304915e58c3cc90de3706df#file-initialize_volume-sh): creates a docker volume named `geoserver_nfs_data`. Downloads the sample datadir file if it's not present in the current directory.
* [compose.yml](https://gist.github.com/groldan/702970e69d18cef763a875ac21fc5e28#file-compose-yml): GeoServer docker composition to load the catalog from an NFS share using geoserver_nfs_data volume

So just download both files and run
```shell
chmod +x ./initialize_volume.sh
./initialize_volume.sh
docker compose up -d
```
When ready GeoServer will be available at http://localhost:8080/geoserver/