@groldan 2025-03-06
> Currently failing on `app-schema-test` but I've no clue what may be causing it. Looking into it

The app schema tests use a legacy data directory and the new loader was [missing a check for it](https://github.com/geoserver/geoserver/compare/315b23474cf8fbd11ccb6416647ce8d0cdfe38e3..5ae48370869baa26570981089cf08e2588c7fc42).