@aaime 2025-03-10
I've tested this latest version, it does not lockup anymore. However, at the same time, I don't see a speed-up anymore, when data directories on the local disk, at least. In both cases the dirs are on a SSD and I've dropped the file system caches before running the bench:


Data dir   | Workspaces | Layers | Current loader | New Loader |
--------------|-------------------|-----------|-----------------------|-------------------|
LargeWS | 1000              | 16k       |  120 sec             | 112 sec         |
LargeLayer| 1                  |  1k        | 23 sec                | 23 sec           |

If you consider the current loader is logging every single layer, feature type, store and resource loaded, and the new one doesn't, those 10 seconds in the large case are going to be even less.
By the way, why was the logging modified? I can see it's limiting performance , but one just needs to set up the logging level to "production" and the issue is gone. For smaller data directories having the logs on is useful.

I've also tried the LargeWS on a Samba share (disk attached to the ADSL router), and got the following timings, but in both cases I stopped the startup after 10 minutes, only a small portion of the catalog was loaded by then (I changed the logging of catalog objects being loaded from FINE to CONFIG to see what was happening).

Can you double check there is any speedup left by using only this PR (leaving out the others about encryption and the like)? Maybe the speedup is visible only with a NFS server.