@groldan 2025-03-11
Hi Andrea,

You're right; as of #ec7a87ac4ec649b62045778ac6ebcf8a7b351666, the performance improvement is not great.
There was a reason for that but it's **fixed** for good now. I'll explain below.

Meanwhile, I've added commit #26c46e4201714fa9625c20645cc3c110ed42e547 to align the logging of the new loader with the current loader (i.e., `config` level log entry for each object added to the catalog, so it’s actually a bit more detailed than the current loader).

# TL;DR

The updated `DataDirectoryGeoServerLoader` (as of this last commit) delivers **faster and safer** catalog loading by parsing in parallel with a work-stealing thread pool and adding sequentially to avoid deadlocks. [GEOS-11758] initially reduced performance gains by prioritizing safety—shifting parsing of Stores, Layers, and LayerGroups to the main thread to prevent Spring-related deadlocks—costing some parallelism. This is now fixed by restoring parallel parsing while ensuring sequential adds for unsafe types, achieving up to 3x faster startup times for large catalogs without compromising reliability.

Data dir (layers)  | Current | New (latest commit) | Improvement |
-|-|-|-|
**16K**| 8.8s | 5.5s | 1.6x |
**100K** | 85.2s | 28.0s | 3x |
**1M** | 792s | 276s | 2.8x |

---

## Test Platform
* OpenJDK Runtime Environment Temurin-11.0.26+4 (build 11.0.26+4)
* GeoServer JVM flags: `-Xmx9G`
* Macbook Pro M3 Max
* 1TB SSD, APFS

## Data Directories:
* [large_shapefiles.tar.bz2](https://www.dropbox.com/scl/fi/ip3x82pdzui1rxnvwsqmb/large_shapefiles.tar.bz2?rlkey=7uk600hlcp9z6tdbsdka86dpl&st=bkz2knvy&dl=1):
  * 1K workspaces
  * 16K layers
* [100K_layers.tar.bz2](https://s3.us-east-1.amazonaws.com/geoserver.datadirectories/100K_layers.tar.bz2) (7.5MB)
  * 100 workspaces
  * 100K layers
* [1M_layers.tar.bz2](https://s3.us-east-1.amazonaws.com/geoserver.datadirectories/1M_layers.tar.bz2) (63.9MB)
  * 1K workspaces
  * 1M layers

## Test Scenarios:

Time to load the Catalog with `DEFAULT_LOGGING` logging profile. Timing taken from log entry `Read Catalog in XXX s, ...`

* **Current**:
  * `DefaultGeoServerLoader`
  * Strategy: 
    * Loads file contents in parallel using fixed thread pool, and adds them to a blocking queue.
    * Parses to `CatalogInfo` and adds them to the `Catalog` in the main thread (queue consumer).
    * This is **safe** but **slow**(ish) in terms of potential deadlocks due to GeoServerExtensions being called during the loading process.
* **New, queued adds**:
  * `DataDirectoryGeoServerLoader`
  * This is the current status of the community module, observing the behaviour mentioned by Andrea, at ec7a87ac4ec649b62045778ac6ebcf8a7b351666
  * Strategy:
    * Loads file contents in work-stealing thread pool, and adds them to a blocking queue.
    * Parses to `CatalogInfo` and adds them to the `Catalog` in the main thread (queue consumer).
    * Potentially offloads I/O (file reads) to additional `ForkJoinPool` threads to prevent blocking on the worker threads.
    * This is **safe** but **slow**(ish) in terms of potential deadlocks due to GeoServerExtensions being called during the loading process.
* **New, parallel parsing and adding**:
  * `DataDirectoryGeoServerLoader`
  * This is how the community module behaved before [[GEOS-11758] datadir-catalog-loader: Improve reliability and code quality](https://github.com/geoserver/geoserver/pull/8385)
  * Strategy:
    * Loads, parses, and adds `CatalogInfo`s to the catalog in work-stealing thread pool.
    * Potentially offloads I/O (file reads) to additional `ForkJoinPool` threads to prevent blocking on the worker threads.
    * This is **fast** but **unsafe**, for it produced **deadlocks** on some data directories when:
      * `XStreamPersister` ended up calling `GeoServerExtensions.bean()/extensions()` concurrently during startup. Spring holds a lock in the main thread for the singleton beans registry. `GeoServerExtensions` forces Spring resolving singletons outside its original plan, since it doesn't follow spring dependency injection practices, bypassing the container’s controlled lifecycle and dependency resolution mechanisms.
      * GeoTools factory lookups are performed concurrently when they're not yet initialized.
    * Workarounds were implemented, like forcing initializing the CRS subsystem, but they turned out not being enough.
    * Most importantly, [GEOS-11758] was executed largely because despite having eliminated the deadlocks in all the known data directories used, still couldn't be sure they wouldn't happen on other edge cases.
* **New, parallel parsing, sequential adds**:
  * `DataDirectoryGeoServerLoader`
  * This is the winner approach, **fast and safe**.
  * Strategy:
    * Loads **and** parses `CatalogInfo`s to the catalog in work-stealing thread pool.
    * Potentially offloads I/O (file reads) to additional `ForkJoinPool` threads to prevent blocking on the worker threads.
    * The `XStreamPersister` **does not resolve ResolvingProxy** references.
    * There are safe and unsafe `CatalogInfo` types in terms of when to resolve proxy references and threading:
      * Safe types are added to the catalog while in the worker threads (workspaces, namespaces, and styles)
      * Unsafe types (stores, resources, and published infos) are added to a queue once parsed, and their proxies resolved in the main (consumer) thread. These ones have the potential of resulting in calls to `GeoServerExtensions.bean/extensions()` and/or `ResourcePool.getDataStore()/getFeatureType()` while `Catalog.add()` validates them.
  

## Test Results

Data dir   | Current | New, queued adds | New, parallel parsing and adding | New, parallel parsing, sequential adds |
---|---|---|---|---|
**large_shapefiles**| 8.8s | 7.5s | 5.4s | 5.5s |
**100K_layers** | 1.42m | 48.3s | 29.9s | 28.0s |
**1M_layers** | 13.2m | *N/A* | *N/A* | 4.6m |

> SMB (or NFS for that matter), over a slow connection for this amount of small files just doesn't make sense. I couldn't find the time to run the tests against a 10Gbps network. But the **large_shapefiles** test scenario on SMB/1Gbps results in `19 secs` for the new loader vs. `45 secs` for the current one.

---

# Startup logs per scenario:

## large_shapefiles:

* **Current**

  ```
  10 Mar 16:23:38 CONFIG [org.geoserver] - Read Catalog in 8.828 s: workspaces: 1,008, namespaces: 1,008, styles: 27, stores: 2,011, resources: 16,025, layers: 16,025, layer groups: 4.
  SEVERE: GeoServer startup complete in 13.0s
  ```

* **New, queued adds**:

  ```
  10 Mar 16:25:52 CONFIG [org.geoserver] - Read Catalog in 7.563 s: workspaces: 1,008, namespaces: 1,008, styles: 27, stores: 2,011, resources: 16,025, layers: 16,025, layer groups: 4.
  SEVERE: GeoServer startup complete in 11.531s
  ```

* **New, parallel parsing and adding**:

  ```
  10 Mar 16:24:42 CONFIG [org.geoserver] - Read Catalog in 5.472 s: workspaces: 1,008, namespaces: 1,008, styles: 27, stores: 2,011, resources: 16,025, layers: 16,025, layer groups: 4.
  SEVERE: GeoServer startup complete in 9.325s
  ```

* **New, parallel parsing, sequential adds**:

  ```
  10 Mar 19:14:29 CONFIG [org.geoserver] - Read Catalog in 5.543 s: workspaces: 1,008, namespaces: 1,008, styles: 27, stores: 2,011, resources: 16,025, layers: 16,025, layer groups: 4.
  SEVERE: GeoServer startup complete in 9.396s
  ```

## 100K_layers:

* **Current**

  ```
  10 Mar 20:13:06 CONFIG [org.geoserver] - Read Catalog in 1.425 min: workspaces: 107, namespaces: 107, styles: 21, stores: 100,009, resources: 100,019, layers: 100,019, layer groups: 3.
  SEVERE: GeoServer startup complete in 89.357s
  ```

* **New, queued adds**:

  ```
  10 Mar 16:09:50 CONFIG [org.geoserver] - Read Catalog in 48.34 s: workspaces: 107, namespaces: 107, styles: 21, stores: 100,009, resources: 100,019, layers: 100,019, layer groups: 3.
  SEVERE: GeoServer startup complete in 55.616s
  ```

* **New, parallel parsing and adding**:

  ```
  10 Mar 16:17:06 CONFIG [org.geoserver] - Read Catalog in 28.94 s: workspaces: 107, namespaces: 107, styles: 21, stores: 100,009, resources: 100,019, layers: 100,019, layer groups: 3.
  SEVERE: GeoServer startup complete in 35.977s
  ```

* **New, parallel parsing, sequential adds**:

  ```
  10 Mar 19:12:55 CONFIG [org.geoserver] - Read Catalog in 28.05 s: workspaces: 107, namespaces: 107, styles: 21, stores: 100,009, resources: 100,019, layers: 100,019, layer groups: 3.
  SEVERE: GeoServer startup complete in 34.732s
  ```


## 1M_layers:

* **Current**

  ```
  10 Mar 21:02:58 CONFIG [org.geoserver] - Read Catalog in 13.27 min: workspaces: 1,007, namespaces: 1,007, styles: 21, stores: 1,000,009, resources: 1,000,019, layers: 1,000,019, layer groups: 3.
  SEVERE: GeoServer startup complete in 805.771s
  ```

* **New, parallel parsing, sequential adds**:

  ```
  10 Mar 21:09:14 CONFIG [org.geoserver] - Read Catalog in 4.681 min: workspaces: 1,007, namespaces: 1,007, styles: 21, stores: 1,000,009, resources: 1,000,019, layers: 1,000,019, layer groups: 3.
  SEVERE: GeoServer startup complete in 287.678s  
  ```